Fix bugs on the blocks listing pages dropdown menus
