+++ public/js/controllers/pagenavigation_controller.js
@@ -4,12 +4,12 @@ import Url from 'url-parse'
 
 export default class extends Controller {
   static get targets () {
-    return ['pagesize', 'votestatus']
+    return ['pagesize', 'votestatus', 'listview']
   }
 
   setPageSize () {
-    let url = Url(window.location.href)
-    let q = Url.qs.parse(url.query)
+    var url = Url(window.location.href)
+    var q = Url.qs.parse(url.query)
     q['offset'] = this.pagesizeTarget.dataset.offset
     q['rows'] = this.pagesizeTarget.selectedOptions[0].value
     url.set('query', q)
@@ -17,10 +17,21 @@ export default class extends Controller {
   }
 
   setFilterbyVoteStatus () {
-    let url = Url(window.location.href)
-    let q = {}
+    var url = Url(window.location.href)
+    var q = {}
     q['byvotestatus'] = this.votestatusTarget.selectedOptions[0].value
     url.set('query', q)
     Turbolinks.visit(url.toString())
+  }
+
+  setListView () {
+    var url = Url(window.location.href, true)
+    var newPeriod = this.listviewTarget.selectedOptions[0].value
+    if (url.pathname !== newPeriod) {
+      delete url.query.offset
+      delete url.query.rows
+    }
+    url.set('pathname', newPeriod)
+    Turbolinks.visit(url.href)
   }
 }
